internal: split up lib.rs
